// backend/src/apis/import/entity-type.enum.ts
export enum ImportEntityType {
  ASSETS = 'immobili',
  ASSET_AGGREGATORS = 'aggregati-immobili',
  UTILITY_AGGREGATORS = 'aggregati-utenze',
  BUDGET_CHAPTERS = 'capitoli-di-spesa',
  SUPPLIERS = 'fornitori',
  UTILIZERS = 'utilizzatori',
  UTILIZER_GRANTS = 'concessioni',
  UTILITIES = 'utenze',
  INVOICES = 'fatture',
}